soma-223/feat - add info type
