+++ pages/login.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react/function-component-definition */
 import { ReactElement } from 'react';
 
+import Layout from '@templates/Layout';
 import KakaoLoginBtn from 'components/Login/atoms/KakaoLoginBtn';
-import Layout from 'components/shared/templates/Layout';
 
 import { NextPageWithLayout } from './_app';
